refactored wallet create dialogs, add folder added
