+++ app/components/wallet/add/WalletRestoreOptionsDialog.js
@@ -2,9 +2,9 @@
 import React, { Component } from 'react';
 import { observer } from 'mobx-react';
 import { defineMessages, intlShape } from 'react-intl';
-import DialogCloseButton from '../widgets/DialogCloseButton';
-import Dialog from '../widgets/Dialog';
-import styles from './WalletCreateDialog.scss';
+import DialogCloseButton from '../../widgets/DialogCloseButton';
+import Dialog from '../../widgets/Dialog';
+import styles from './WalletCreateOptions.scss';
 import WalletTypeItem from './WalletTypeItem';
 
 const messages = defineMessages({
@@ -31,6 +31,14 @@ const messages = defineMessages({
   walletPaperDescription: {
     id: 'wallet.create.type.paper.description',
     defaultMessage: '!!!Allows the generation of an offline wallet that should be printed. The process involves setting up an optional password for extra security and then printing a pdf that contains the mnemonics (for later recovery) and some Cardano addresses onto a paper.',
+  },
+  walletTabsRecommended: {
+    id: 'wallet.create.tabs.recommended.title',
+    defaultMessage: '!!!Recommended',
+  },
+  walletTabsAdvanced: {
+    id: 'wallet.create.tabs.advanced.title',
+    defaultMessage: '!!!Advanced',
   },
 });
 
@@ -59,8 +67,20 @@ export default class WalletRestoreOptionsDialog extends Component<Props> {
         closeButton={<DialogCloseButton />}
         classicTheme={classicTheme}
       >
+        {/* <div className={styles.tabs}>
+          <button
+            type="button"
+            className={`${styles.tabsLink} ${styles.active}`}>
+            {intl.formatMessage(messages.walletTabsRecommended)}
+          </button>
+          <button
+            type="button"
+            className={styles.tabsLink}>
+            {intl.formatMessage(messages.walletTabsAdvanced)}
+          </button>
+        </div> */}
         <div className={styles.tabsContent}>
-          <ul className={styles.heroWalletTypeList}>
+          <ul className={styles.WalletTypeList}>
             <WalletTypeItem
               action={onRestore}
               type="mnemonic"
