+++ app/components/wallet/add/WalletTypeItem.scss
@@ -74,6 +74,7 @@
     transition: max-height 300ms linear;
   }
 }
+// LEARN MORE BUTTON
 
 .moreBtn {
   color: #ADAEB6;
@@ -118,27 +119,27 @@
   .walletTypeImg {
 
     &.mnemonic {
-      background-image: url(../../assets/images/wallets/modern/mnemonic_wallet.svg);
+      background-image: url(../../../assets/images/wallets/modern/mnemonic_wallet.svg);
     }
   
     &.master {
-      background-image: url(../../assets/images/wallets/modern/master_key.svg);
+      background-image: url(../../../assets/images/wallets/modern/master_key.svg);
     }
   
     &.paper {
-      background-image: url(../../assets/images/wallets/modern/paper.wallet.inline.svg);
+      background-image: url(../../../assets/images/wallets/modern/paper.wallet.inline.svg);
     }
   
     &.business {
-      background-image: url(../../assets/images/wallets/modern/business_wallet.svg);
+      background-image: url(../../../assets/images/wallets/modern/business_wallet.svg);
     }
 
     &.trezor {
-      background-image: url(../../assets/images/wallets/modern/trezor.inline.svg);
+      background-image: url(../../../assets/images/wallets/modern/trezor.inline.svg);
     }
 
     &.ledger {
-      background-image: url(../../assets/images/wallets/modern/ledger.inline.svg);
+      background-image: url(../../../assets/images/wallets/modern/ledger.inline.svg);
     }
   }
 
@@ -151,33 +152,32 @@
   }
 }
 
-
 :global(.YoroiClassic) {
 
   .walletTypeImg {
 
     &.mnemonic {
-      background-image: url(../../assets/images/wallets/classic/mnemonimic_words_wallet.inline.svg);
+      background-image: url(../../../assets/images/wallets/classic/mnemonimic_words_wallet.inline.svg);
     }
   
     &.master {
-      background-image: url(../../assets/images/wallets/classic/master_key.inline.svg);
+      background-image: url(../../../assets/images/wallets/classic/master_key.inline.svg);
     }
   
     &.paper {
-      background-image: url(../../assets/images/wallets/classic/paper_wallet.inline.svg);
+      background-image: url(../../../assets/images/wallets/classic/paper_wallet.inline.svg);
     }
   
     &.business {
-      background-image: url(../../assets/images/wallets/classic/business_wallet.inline.svg);
+      background-image: url(../../../assets/images/wallets/classic/business_wallet.inline.svg);
     }
 
     &.trezor {
-      background-image: url(../../assets/images/wallets/classic/trezor_wallet.inline.svg);
+      background-image: url(../../../assets/images/wallets/classic/trezor_wallet.inline.svg);
     }
 
     &.ledger {
-      background-image: url(../../assets/images/wallets/classic/ledger_wallet.inline.svg);
+      background-image: url(../../../assets/images/wallets/classic/ledger_wallet.inline.svg);
     }
   }
